+++ spec/sound/win32/win32_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 if OS.windows?
-  describe Sound::Win32::Sound do
+  describe Sound::Win32 do
     
   end
 end
